implement the modal for user credentials on ordering menu

+++ src/app/pages/shopping-card/shopping-card.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, ViewChild, ElementRef, OnInit } from '@angular/core';
 import { LocalStorageService } from 'ng2-webstorage';
 
 import { ApiCallsService } from '../../providers/api-calls.service';
@@ -10,6 +10,8 @@ import { NotificationService } from '../../providers/notification.service';
   styleUrls: ['./shopping-card.component.scss']
 })
 export class ShoppingCardComponent implements OnInit {
+  @ViewChild('staticModal') public staticModal: ElementRef;
+
   public selectedMenuList: any;
   public menuOrder = {};
   public totalPrice  = 0;
@@ -17,7 +19,15 @@ export class ShoppingCardComponent implements OnInit {
     private localSt: LocalStorageService,
     private apiCallsService: ApiCallsService,
     private notificationService: NotificationService
-  ) { }
+  ) {
+    this.menuOrder['customer'] = {
+      name: null,
+      email: null,
+      address: null,
+      telephone: null,
+      comment: null
+    };
+  }
 
   ngOnInit() {
     this.getSelectedMenuList();
@@ -26,14 +36,12 @@ export class ShoppingCardComponent implements OnInit {
 
   public getSelectedMenuList() {
     this.selectedMenuList = this.localSt.retrieve('shop-card-list');
-    console.log('******** selectedMenuList', this.selectedMenuList);
   }
 
   public deleteMenu(item) {
     const index = this.selectedMenuList.indexOf(item);
     if (index > -1) {
       this.selectedMenuList.splice(index, 1);
-      console.log('******** selectedMenuList', this.selectedMenuList);
       this.localSt.store('shop-card-list', this.selectedMenuList);
     }
     this.calculateTotalPrice();
@@ -46,26 +54,29 @@ export class ShoppingCardComponent implements OnInit {
   }
 
   public orderMenus() {
-    this.menuOrder['customer'] = {
-      name: 'hamza jridi',
-      address: 'Tunis',
-      tel: '22428402',
-      email: 'hamza@me.com'
-    };
+    if (!this.menuOrder['customer']['email'] && !this.menuOrder['customer']['telephone']) {
+      return;
+    }
+    // this.menuOrder['customer'] = {
+    //   name: 'hamza jridi',
+    //   address: 'Tunis',
+    //   tel: '22428402',
+    //   email: 'hamza@me.com'
+    // };
     this.menuOrder['order'] =  this.calculatePrices(this.selectedMenuList, false);
     this.calculateTotalPrice();
+
     this.apiCallsService.postData('order', this.menuOrder).subscribe(
       (data) => {
-        console.log('*********** data', data);
         this.notificationService.notification.next({
           msgType: 'success',
           msgTitle: 'Order Confirmation',
           msgContent: 'Your order has been submitted, we thank you for chosing us',
         });
         this.deleteAllMenus();
+        this.staticModal.hide();
       },
       (error) => {
-        console.log('*********** error', error);
         this.notificationService.notification.next({
           msgType: 'success',
           msgTitle: 'Order Error',
@@ -79,16 +90,13 @@ export class ShoppingCardComponent implements OnInit {
     this.totalPrice = this.menuOrder['totalPrice'];
   }
   public getItemTotalPrice(item) {
-    console.log('********** item.quantity', item.quantity);
     item['totalPrice'] = item['price'] * item['quantity'];
-    console.log('******** item[totalPrice]', item['totalPrice']);
     this.calculateTotalPrice();
   }
 
   public calculatePrices(orderList, price) {
     let totalPrice = 0;
     for (let item of orderList) {
-      console.log('********* item', item);
       item['menuItem']['totalPrice'] = item['menuItem']['price'] * item['menuItem']['quantity'];
       totalPrice += item['menuItem']['totalPrice'];
     }
